requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
beartype==0.22.6
bleach==6.3.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
datasets==4.4.2
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.1
expecttest==0.3.0
fastai==2.8.5
fastcore==1.8.16
fastdownload==0.0.7
fastjsonschema==2.21.2
fastprogress==1.0.3
fastrelease==0.1.17
fasttransform==0.0.2
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
ghapi==1.0.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-server==1.15.6
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab==3.3.4
jupyterlab_server==2.16.3
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
nbclassic==0.5.6
nbconvert==5.6.1
nbdev==1.1.21
nbformat==5.10.4
nbverbose==0.0.9
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
notebook==6.5.4
notebook_shim==0.2.4
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
plum-dispatch==2.6.0
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rouge_score==0.1.2
rpds-py==0.29.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
sentencepiece==0.2.1
seqeval==1.2.2
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
testpath==0.6.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
#torch==2.4.1
torch-directml==0.2.5.dev240914
torchvision==0.19.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typedispatch==1.1.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==2.0.1
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0